refactor(loaders): split problem loader into a package and tighten contest/API plumbing
- Split monolithic problem.py into a problem/ package: dispatch in __init__, parsing in _parse, list resolution in _resolve, format loaders in domjudge.py and polygon.py.
- Resolve team username collisions deterministically in contest loader by re-hashing with a collision suffix until a free handle is found (10000-bucket hash had ~14% birthday collision risk at 55 teams).
- ContestStateComparator now uses team.username directly instead of recomputing the hash (which skipped collision probing).
- Drop unused cache_ttl/enable_cache plumbing from APIClientFactory; clients are constructed with just base_url/username/password.
